requirement.txt
discord.py
jinja2
cairosvg
cairo
